+++ src/main/java/de/samply/core/ExporterParameters.java
@@ -13,5 +13,6 @@ public record ExporterParameters(
     String queryLabel,
     String queryDescription,
     String queryContactId,
+    String queryContext,
     LocalDate queryExpirationDate,
     Format outputFormat) {}