+++ src/test/java/de/samply/core/ExporterCoreTest.java
@@ -52,7 +52,7 @@ class ExporterCoreTest {
     @Test
     void retrieveByQueryId() throws ExporterCoreException {
         ExporterParameters exporterParameters = new ExporterParameters(null, "Patient",
-                converterTemplateId, null, null, Format.FHIR_SEARCH, null, null, null, null, Format.CSV);
+                converterTemplateId, null, null, Format.FHIR_SEARCH, null, null, null, null, null, Format.CSV);
         ExporterCoreParameters exporterCoreParameters = exporterCore.extractParameters(
                 exporterParameters);
         Flux<Path> resultFlux = exporterCore.retrieveQuery(exporterCoreParameters, tokenContext);